[CE-152] Added document for an all-in-one deployment
Currently there is no specific document on how to stand up
a fabric network on a single node. This patch set adds the
missing document so that developers and users can follow
the steps to use an existing machine to setup all in one
fabric network. This patch set also fixes a small bug when
a machine may have comments in /etc/resolv.conf file which
will cause the problem for nameserver misconfiguration.

Change-Id: Icb5091c1ed03662c76cac4e3ed62de0918d92be6
Signed-off-by: tongliofcary <litong01@us.ibm.com>