[CE-195]build target change caused containers not found
abric build target docker has been reduced not to build
zookeeper, kafka and couchdb images. this change breaks
the ansible agent when docker is a build target but the
images are missing. We need to change this to allow compose
and k8s draw specific zookeeper, kafka, fabric-ca and
couchdb images from docker hub.

Change-Id: I990e906b82922e46d0bb7da1cb3c626c8be24da6
Signed-off-by: tongliofcary <litong01@us.ibm.com>