[CE-488] Fabric Chrat for Cello
This is a K8S Chart for Fabric deployment, the Chart is placed
at cello/thirdparty/helm/ directory, it is independent from
Cello which can be launched without Cello is running.

The purpose of this Chart is to provide a method for user to deploy
Fabric on Kubernetes easily and offer a place for user to give
their feedback, finally, we will replace the current kubernetes
agent with Helm/Chart.

CE-488 #done

Change-Id: Ic74ecdcd2200b1781544247a4989972e62f0e737
Signed-off-by: luke <jiahaochen1993@gmail.com>